🔒 Improve public bin packing interface
